WIP: split some component definitions into individual files.
