+++ src/components/box/index.d.ts
@@ -1,0 +1,5 @@
+/// <reference path="../../../index.d.ts" />
+
+import { BulmaComponent } from '../../../index';
+
+export const Box: BulmaComponent<{}, 'div'>;
