+++ src/components/container/index.d.ts
@@ -1,0 +1,8 @@
+import { Breakpoint, BulmaComponent } from '../../../index';
+
+interface ContainerProps {
+  fluid?: boolean;
+  breakpoint?: Breakpoint;
+}
+
+export const Container: BulmaComponent<ContainerProps, 'div'>;
