+++ src/components/content/index.d.ts
@@ -1,0 +1,7 @@
+import { BulmaComponent, Size } from '../../../index';
+
+interface ContentProps {
+  size?: Size;
+}
+
+export const Content: BulmaComponent<ContentProps, 'div'>;
