+++ tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+    "module": "commonjs",
+    "target": "es5",
+    "sourceMap": true,
+    "baseUrl": ".",
+    "typeRoots": ["."]
+  },
+  "exclude": [
+    "node_modules"
+  ]
+}
